+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: ide-ruit <ide-ruit@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/18 14:09:30 by ide-ruit          #+#    #+#             */
-/*   Updated: 2024/02/10 19:10:08 by ide-ruit         ###   ########.fr       */
+/*   Updated: 2024/02/14 20:40:32 by ide-ruit         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *str1, const char *str2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while ((str1[i] != '\0' && str2[i] != '\0') && i <= n)
